populating DB
add getting exchange data from csv file from bundesbank, refactoring

+++ src/main/java/com/crewmeister/cmcodingchallenge/external/CurrencyQuote.java
@@ -3,7 +3,6 @@ package com.crewmeister.cmcodingchallenge.external;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -11,5 +10,5 @@ public class CurrencyQuote {
 
     String currencyCode;
 
-    BigDecimal exchangeRate;
+    Double exchangeRate;
 }
